create api and separate out wftools
