+++ pyqmc/recipes.py
@@ -1,8 +1,7 @@
-import pyqmc
-import pyqmc.obdm
+from pyqmc import obdm
+import pyqmc.api as pyq
 import numpy as np
 import h5py
-import pyqmc.reblock
 import scipy.stats
 import pandas as pd
 import copy
@@ -27,25 +26,25 @@ def OPTIMIZE(
 
     target_root = 0
     if ci_checkfile is None:
-        mol, mf = pyqmc.recover_pyscf(dft_checkfile)
+        mol, mf = pyq.recover_pyscf(dft_checkfile)
         mc = None
     else:
-        mol, mf, mc = pyqmc.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
+        mol, mf, mc = pyq.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
         mc.ci = mc.ci[target_root]
 
     if S is not None:
-        mol = pyqmc.get_supercell(mol, np.asarray(S))
+        mol = pyq.get_supercell(mol, np.asarray(S))
 
-    wf, to_opt = pyqmc.generate_wf(
+    wf, to_opt = pyq.generate_wf(
         mol, mf, mc=mc, jastrow_kws=jastrow_kws, slater_kws=slater_kws
     )
     if start_from is not None:
-        pyqmc.read_wf(wf, start_from)
+        pyq.read_wf(wf, start_from)
 
-    configs = pyqmc.initial_guess(mol, nconfig)
-    acc = pyqmc.gradient_generator(mol, wf, to_opt)
+    configs = pyq.initial_guess(mol, nconfig)
+    acc = pyq.gradient_generator(mol, wf, to_opt)
     if anchors is None:
-        pyqmc.line_minimization(
+        pyq.line_minimization(
             wf,
             configs,
             acc,
@@ -56,9 +55,9 @@ def OPTIMIZE(
             **linemin_kws
         )
     else:
-        wfs = [pyqmc.read_wf(copy.deepcopy(wf), a) for a in anchors]
+        wfs = [pyq.read_wf(copy.deepcopy(wf), a) for a in anchors]
         wfs.append(wf)
-        pyqmc.optimize_orthogonal(
+        pyq.optimize_orthogonal(
             wfs,
             configs,
             acc,
@@ -83,7 +82,7 @@ def generate_accumulators(mol, mf, energy=True, rdm1=False, extra_accumulators=N
     if energy:
         if "energy" in acc.keys():
             raise Exception("Found energy in extra_accumulators and energy is True")
-        acc["energy"] = pyqmc.EnergyAccumulator(mol)
+        acc["energy"] = pyq.EnergyAccumulator(mol)
     if rdm1:
         if len(mf.mo_coeff.shape) == 2:
             mo_coeff = [mf.mo_coeff, mf.mo_coeff]
@@ -93,12 +92,8 @@ def generate_accumulators(mol, mf, energy=True, rdm1=False, extra_accumulators=N
             raise Exception(
                 "Found rdm1_up or rdm1_down in extra_accumulators and rdm1 is True"
             )
-        acc["rdm1_up"] = (
-            pyqmc.obdm.OBDMAccumulator(mol, orb_coeff=mo_coeff[0], spin=0),
-        )
-        acc["rdm1_down"] = (
-            pyqmc.obdm.OBDMAccumulator(mol, orb_coeff=mo_coeff[1], spin=1),
-        )
+        acc["rdm1_up"] = (obdm.OBDMAccumulator(mol, orb_coeff=mo_coeff[0], spin=0),)
+        acc["rdm1_down"] = (obdm.OBDMAccumulator(mol, orb_coeff=mo_coeff[1], spin=1),)
 
     return acc
 
@@ -122,29 +117,29 @@ def VMC(
 
     target_root = 0
     if ci_checkfile is None:
-        mol, mf = pyqmc.recover_pyscf(dft_checkfile)
+        mol, mf = pyq.recover_pyscf(dft_checkfile)
         mc = None
     else:
-        mol, mf, mc = pyqmc.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
+        mol, mf, mc = pyq.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
         mc.ci = mc.ci[target_root]
 
     if S is not None:
         print("S", S)
-        mol = pyqmc.get_supercell(mol, np.asarray(S))
+        mol = pyq.get_supercell(mol, np.asarray(S))
 
     if accumulators is None:
         accumulators = {}
 
-    wf, _ = pyqmc.generate_wf(
+    wf, _ = pyq.generate_wf(
         mol, mf, mc=mc, jastrow_kws=jastrow_kws, slater_kws=slater_kws
     )
 
     if start_from is not None:
-        pyqmc.read_wf(wf, start_from)
+        pyq.read_wf(wf, start_from)
 
-    configs = pyqmc.initial_guess(mol, nconfig)
+    configs = pyq.initial_guess(mol, nconfig)
 
-    pyqmc.vmc(
+    pyq.vmc(
         wf,
         configs,
         accumulators=generate_accumulators(mol, mf, **accumulators),
@@ -175,25 +170,27 @@ def DMC(
 
     target_root = 0
     if ci_checkfile is None:
-        mol, mf = pyqmc.recover_pyscf(dft_checkfile)
+        mol, mf = pyq.recover_pyscf(dft_checkfile)
         mc = None
     else:
-        mol, mf, mc = pyqmc.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
+        mol, mf, mc = pyq.recover_pyscf(dft_checkfile, ci_checkfile=ci_checkfile)
         mc.ci = mc.ci[target_root]
 
     if S is not None:
-        mol = pyqmc.get_supercell(mol, np.asarray(S))
+        mol = pyq.get_supercell(mol, np.asarray(S))
     if accumulators is None:
         accumulators = {}
 
-    wf, _ = pyqmc.generate_wf(mol, mf, mc=mc, jastrow_kws=jastrow_kws, slater_kws=slater_kws)
+    wf, _ = pyq.generate_wf(
+        mol, mf, mc=mc, jastrow_kws=jastrow_kws, slater_kws=slater_kws
+    )
 
     if start_from is not None:
-        pyqmc.read_wf(wf, start_from)
+        pyq.read_wf(wf, start_from)
 
-    configs = pyqmc.initial_guess(mol, nconfig)
+    configs = pyq.initial_guess(mol, nconfig)
 
-    pyqmc.rundmc(
+    pyq.rundmc(
         wf,
         configs,
         accumulators=generate_accumulators(mol, mf, **accumulators),
@@ -222,7 +219,7 @@ def read_mc_output(fname, warmup=5, reblock=16):
     with h5py.File(fname) as f:
         for k in f.keys():
             if "energy" in k:
-                vals = pyqmc.reblock.reblock(f[k][warmup:], reblock)
+                vals = pyq.avg_reblock(f[k][warmup:], reblock)
                 ret[k] = np.mean(vals)
                 ret[k + "_err"] = scipy.stats.sem(vals)
     return ret
